Merge main branch changes while preserving the guess verification feature:
- Keep 3 public signals (commitment, isCorrect, guess) in verifier
- Verify proof's guess matches the submitted challenge guess
- Use uint256 types for consistency with main branch style
- Keep IC3 constants for the third public signal verification

+++ test/invariants/GuessGameInvariants.t.sol
@@ -11,56 +11,49 @@ contract GuessGameInvariants is Test {
     GuessGame public game;
     Groth16Verifier public verifier;
     GuessGameHandler public handler;
-    
+
     function setUp() public {
         // Deploy contracts
         verifier = new Groth16Verifier();
         game = new GuessGame(address(verifier));
         handler = new GuessGameHandler(game, verifier);
-        
+
         // Configure invariant test
         targetContract(address(handler));
-        
+
         // Only call handler functions
         bytes4[] memory selectors = new bytes4[](4);
         selectors[0] = GuessGameHandler.createPuzzle.selector;
         selectors[1] = GuessGameHandler.submitGuess.selector;
         selectors[2] = GuessGameHandler.respondToChallenge.selector;
         selectors[3] = GuessGameHandler.closePuzzle.selector;
-        
-        targetSelector(FuzzSelector({
-            addr: address(handler),
-            selectors: selectors
-        }));
-        
+
+        targetSelector(FuzzSelector({addr: address(handler), selectors: selectors}));
+
         // Fund handler's actors
-        for (uint i = 0; i < 3; i++) {
+        for (uint256 i = 0; i < 3; i++) {
             vm.deal(address(uint160(0x1000 + i)), 100 ether); // creators
             vm.deal(address(uint160(0x2000 + i)), 100 ether); // guessers
         }
     }
-    
+
     /**
      * @notice Contract balance should always equal sum of all active puzzles' funds
      */
     function invariant_contractBalanceMatchesPuzzleFunds() public view {
         uint256 contractBalance = address(game).balance;
         uint256 expectedBalance = handler.sumActivePuzzleFunds();
-        
-        assertEq(
-            contractBalance,
-            expectedBalance,
-            "Contract balance doesn't match sum of puzzle funds"
-        );
+
+        assertEq(contractBalance, expectedBalance, "Contract balance doesn't match sum of puzzle funds");
     }
-    
+
     /**
      * @notice Puzzle and challenge IDs should only increase monotonically
      */
     function invariant_monotonicIds() public view {
         uint256 puzzleCount = game.puzzleCount();
         uint256 challengeCount = game.challengeCount();
-        
+
         // Check each puzzle ID exists in sequence
         for (uint256 i = 0; i < puzzleCount; i++) {
             IGuessGame.Puzzle memory puzzle = game.getPuzzle(i);
@@ -70,7 +63,7 @@ contract GuessGameInvariants is Test {
                 assert(puzzle.bounty >= 0.001 ether);
             }
         }
-        
+
         // Check each challenge ID exists in sequence
         for (uint256 i = 0; i < challengeCount; i++) {
             IGuessGame.Challenge memory challenge = game.getChallenge(i);
@@ -78,20 +71,20 @@ contract GuessGameInvariants is Test {
             assert(challenge.guesser != address(0));
         }
     }
-    
+
     /**
      * @notice Each puzzle's totalStaked should equal sum of all its challenge stakes
      */
     function invariant_totalStakedMatchesChallengeStakes() public view {
         uint256 puzzleCount = game.puzzleCount();
-        
+
         for (uint256 puzzleId = 0; puzzleId < puzzleCount; puzzleId++) {
             IGuessGame.Puzzle memory puzzle = game.getPuzzle(puzzleId);
             if (puzzle.creator == address(0)) continue; // Skip deleted puzzles
-            
+
             uint256 sumOfStakes = 0;
             uint256 challengeCount = game.challengeCount();
-            
+
             // Sum up all stakes for this puzzle
             for (uint256 challengeId = 0; challengeId < challengeCount; challengeId++) {
                 if (game.challengeToPuzzle(challengeId) == puzzleId) {
@@ -99,21 +92,17 @@ contract GuessGameInvariants is Test {
                     sumOfStakes += challenge.stake;
                 }
             }
-            
-            assertEq(
-                puzzle.totalStaked,
-                sumOfStakes,
-                "Puzzle totalStaked doesn't match sum of challenge stakes"
-            );
+
+            assertEq(puzzle.totalStaked, sumOfStakes, "Puzzle totalStaked doesn't match sum of challenge stakes");
         }
     }
-    
+
     /**
      * @notice Solved puzzles should remain solved forever
      */
     function invariant_solvedPuzzlesImmutable() public view {
         uint256 puzzleCount = game.puzzleCount();
-        
+
         for (uint256 i = 0; i < puzzleCount; i++) {
             if (handler.ghostPuzzleSolved(i)) {
                 IGuessGame.Puzzle memory puzzle = game.getPuzzle(i);
@@ -121,19 +110,19 @@ contract GuessGameInvariants is Test {
             }
         }
     }
-    
+
     /**
      * @notice Challenge responses should be immutable
      */
     function invariant_challengeResponsesImmutable() public view {
         uint256 challengeCount = game.challengeCount();
-        
+
         for (uint256 i = 0; i < challengeCount; i++) {
             IGuessGame.Challenge memory challenge = game.getChallenge(i);
             // Check consistency with puzzle state
             uint256 puzzleId = game.challengeToPuzzle(i);
             IGuessGame.Puzzle memory puzzle = game.getPuzzle(puzzleId);
-            
+
             // If puzzle is solved, at least one challenge must be responded with correct guess
             if (puzzle.solved && challenge.responded) {
                 // This challenge might be the winning one
@@ -141,13 +130,13 @@ contract GuessGameInvariants is Test {
             }
         }
     }
-    
+
     /**
      * @notice Growth percent should always be <= 100
      */
     function invariant_growthPercentBounds() public view {
         uint256 puzzleCount = game.puzzleCount();
-        
+
         for (uint256 i = 0; i < puzzleCount; i++) {
             IGuessGame.Puzzle memory puzzle = game.getPuzzle(i);
             if (puzzle.creator != address(0)) {
@@ -155,13 +144,13 @@ contract GuessGameInvariants is Test {
             }
         }
     }
-    
+
     /**
      * @notice All puzzles should have minimum bounty if they exist
      */
     function invariant_minimumBounty() public view {
         uint256 puzzleCount = game.puzzleCount();
-        
+
         for (uint256 i = 0; i < puzzleCount; i++) {
             IGuessGame.Puzzle memory puzzle = game.getPuzzle(i);
             if (puzzle.creator != address(0)) {
@@ -169,22 +158,22 @@ contract GuessGameInvariants is Test {
             }
         }
     }
-    
+
     /**
      * @notice Creator rewards accounting should be consistent
      */
     function invariant_creatorRewardAccounting() public view {
         uint256 puzzleCount = game.puzzleCount();
-        
+
         for (uint256 puzzleId = 0; puzzleId < puzzleCount; puzzleId++) {
             IGuessGame.Puzzle memory puzzle = game.getPuzzle(puzzleId);
             if (puzzle.creator == address(0) || puzzle.solved) continue;
-            
+
             // Calculate expected creator reward from incorrect guesses
             uint256 expectedCreatorReward = 0;
             uint256 expectedBountyGrowth = 0;
             uint256 challengeCount = game.challengeCount();
-            
+
             for (uint256 challengeId = 0; challengeId < challengeCount; challengeId++) {
                 if (game.challengeToPuzzle(challengeId) == puzzleId) {
                     IGuessGame.Challenge memory challenge = game.getChallenge(challengeId);
@@ -196,13 +185,13 @@ contract GuessGameInvariants is Test {
                     }
                 }
             }
-            
+
             // Due to the way we handle responses, we can't perfectly track this
             // But creator reward should never exceed total stakes
             assert(puzzle.creatorReward <= puzzle.totalStaked);
         }
     }
-    
+
     /**
      * @notice No ETH should be created or destroyed
      */
@@ -211,7 +200,7 @@ contract GuessGameInvariants is Test {
         // But we can add additional checks
         uint256 contractBalance = address(game).balance;
         uint256 totalTracked = handler.ghostTotalContractFunds();
-        
+
         // Contract balance should never exceed what we've tracked going in
         assert(contractBalance <= totalTracked);
     }
